+++ src/blood-oxygenation/domain/bloodOxygenation.go
@@ -1,7 +1,7 @@
 package domain
 
 type IBloodOxygenation interface {
-	SaveBloodOxygenation(esp32ID string, timestamp string, ir int32, red int32, spo2 int32) error
+	SaveBloodOxygenation(esp32ID string, timestamp string, ir int32, red int32, spo2 float32) error
 	GetAll() ([]BloodOxygenation, error)
 }
 
@@ -11,10 +11,10 @@ type BloodOxygenation struct {
 	Timestamp string  `json:"tiempo"`
 	IR int32   `json:"ir"`
 	Red int32  `json:"red"`
-	SpO2 int32  `json:"spo2"`
+	SpO2 float32  `json:"spo2"`
 }
 
-func NewBloodOxygenation(esp32ID string, tiempo string, ir int32, red int32, spo2 int32) *BloodOxygenation {
+func NewBloodOxygenation(esp32ID string, tiempo string, ir int32, red int32, spo2 float32) *BloodOxygenation {
 	return &BloodOxygenation{
 		ESP32ID:  esp32ID,
 		Timestamp: tiempo,
@@ -24,7 +24,7 @@ func NewBloodOxygenation(esp32ID string, tiempo string, ir int32, red int32, spo
 	}
 }
 
-func (bt *BloodOxygenation) SetBloodOxygenation(ir int32, red int32, spo2 int32) {
+func (bt *BloodOxygenation) SetBloodOxygenation(ir int32, red int32, spo2 float32) {
 	bt.IR = ir
 	bt.Red = red
 	bt.SpO2 = spo2
